refactor: move to domain
